Switching app engine tests to webtest and consolidating testing utils into AppEngineTestbedCase
