+++ appengine/memcache/guestbook/tests/test_guestbook.py
@@ -14,18 +14,15 @@
 
 # from the app main.py
 from appengine.memcache.guestbook import main
-from tests import DatastoreTestbedCase
+from tests import AppEngineTestbedCase
+import webtest
 
-import webapp2
 
+class TestHandlers(AppEngineTestbedCase):
 
-class TestHandlers(DatastoreTestbedCase):
     def test_hello(self):
-        # Build a request object passing the URI path to be tested.
-        # You can also pass headers, query arguments etc.
-        request = webapp2.Request.blank('/')
-        # Get a response for that request.
-        response = request.get_response(main.app)
+        app = webtest.TestApp(main.app)
+        response = app.get('/')
 
         # Let's check if the response is correct.
         self.assertEqual(response.status_int, 200)
