+++ appengine/ndb/modeling/tests/test_parent_child_models.py
@@ -15,13 +15,11 @@
 """Test classes for code snippet for modeling article."""
 
 from appengine.ndb.modeling import parent_child_models as models
-
 from google.appengine.ext import ndb
+from tests import AppEngineTestbedCase
 
-from tests import DatastoreTestbedCase
 
-
-class ContactTestCase(DatastoreTestbedCase):
+class ContactTestCase(AppEngineTestbedCase):
     """A test case for the Contact model class with KeyProperty."""
     NAME = 'Takashi Matsuo'
