+++ appengine/ndb/modeling/tests/test_relation_model_models.py
@@ -15,13 +15,11 @@
 """Test classes for code snippet for modeling article."""
 
 from appengine.ndb.modeling import relation_model_models as models
-
 from google.appengine.ext import ndb
+from tests import AppEngineTestbedCase
 
-from tests import DatastoreTestbedCase
 
-
-class ContactTestCase(DatastoreTestbedCase):
+class ContactTestCase(AppEngineTestbedCase):
     """A test case for the Contact model with relationship model."""
     def setUp(self):
         """Creates 1 contact and 1 company.
