+++ appengine/ndb/modeling/tests/test_structured_property_models.py
@@ -15,11 +15,10 @@
 """Test classes for code snippet for modeling article."""
 
 from appengine.ndb.modeling import structured_property_models as models
+from tests import AppEngineTestbedCase
 
-from tests import DatastoreTestbedCase
 
-
-class ContactTestCase(DatastoreTestbedCase):
+class ContactTestCase(AppEngineTestbedCase):
     """A test case for the Contact model with StructuredProperty."""
     def setUp(self):
         """Creates one Contact entity with 2 phone numbers."""
