+++ tests/__init__.py
@@ -13,20 +13,20 @@
 #
 
 from .utils import (
+    AppEngineTestbedCase,
     BUCKET_NAME_ENV,
     capture_stdout,
     CloudBaseTest,
-    DatastoreTestbedCase,
     mock_input_answers,
     PROJECT_ID_ENV,
     RESOURCE_PATH)
 
 
 __all__ = [
+    'AppEngineTestbedCase',
     'BUCKET_NAME_ENV',
     'capture_stdout',
     'CloudBaseTest',
-    'DatastoreTestbedCase',
     'mock_input_answers',
     'PROJECT_ID_ENV',
     'RESOURCE_PATH'
