+++ tox.ini
@@ -12,6 +12,7 @@ deps =
     mock
     nose
     coverage
+    webtest
     nose-exclude
 coverargs =
     --with-coverage
